backend/pyproject.toml
[project]
name = "transcribee-backend"
version = "0.1.0"
description = "The backend for the transcribee open-source transcription platform."
authors = [
    { name = "anuejn", email = "jarohabiger@googlemail.com" },
    { name = "pajowu", email = "git@ca.pajowu.de" },
    { name = "phlmn", email = "info@philipp-mandler.com" },
    { name = "vup", email = "robin.ole.heinemann+transcribee@gmail.com" },
]

dependencies = [
    "fastapi>=0.92.0",
    "uvicorn[standard]>=0.20.0",
    "sqlmodel @ git+https://github.com/transcribee/sqlmodel.git@transcribee_main",
    "alembic>=1.9.4",
    "transcribee-proto @ file:///${PROJECT_ROOT}/../proto",
    "python-multipart>=0.0.6",
    "filetype>=1.2.0",
    "websockets>=10.4",
    "python-magic>=0.4.27",
]
requires-python = ">=3.10"
readme = "./README.md"
license = { text = "AGPL-3.0" }


[tool.pdm.dev-dependencies]
dev = [
    "-e file:///${PROJECT_ROOT}/../proto#egg=transcribee-proto",
    "pyyaml>=6.0",
]

[tool.pdm.scripts]
start = "uvicorn transcribee_backend.main:app"
dev = "uvicorn transcribee_backend.main:app --reload"
migrate = "alembic upgrade head"
makemigrations = "alembic revision --autogenerate -m"
create_user = "scripts/create_user.py"
create_worker = "scripts/create_worker.py"
generate_openapi = "python -m scripts.generate_openapi"

[build-system]
requires = ["pdm-pep517>=1.0"]
build-backend = "pdm.pep517.api"
